First pass of refactor for the cpp code.
